Add files via upload
For some bizarre reason, the FIPS file that Kevin used to merge the FIPS codes to county_yearly does not work for the state. The dict reader does not pick up 'State' so I am attempting to use another file with the exact same data to complete the same merge except at the state level.